refactor(utils): split taf decoder helpers
Move TAF token patterns, element decoding, segment parsing, and time formatting into focused modules while keeping taf_decoder compatibility exports stable.

Tested: focused TAF and aviation/weather-client pytest suites; ruff check src tests scripts; pyright; native toast/audio/thread pytest slice; suite-minus-native-thread-files pytest slice

Confidence: high
